fix: explicitly cast delta time to float for position updates in the movement system

+++ ecs/src/systems/MovementSystem.cpp
@@ -8,8 +8,8 @@ void MovementSystem::update(GameEngine::Registry& registry, double dt) {
     auto view = registry.view<component::Position, component::Velocity>();
 
     view.each([dt](component::Position& pos, component::Velocity& vel) {
-        pos.x += vel.vx * dt;
-        pos.y += vel.vy * dt;
+        pos.x += vel.vx * static_cast<float>(dt);
+        pos.y += vel.vy * static_cast<float>(dt);
     });
 }
